requirements.txt
Flask
ultralytics
Pillow
easyocr
numpy
opencv-python